use seconds as world time unit

+++ src/client/NetworkedInput.hpp
@@ -37,9 +37,9 @@ private:
 	struct PlayerHistory {
 		EntityState state;
 		glm::vec3 tgt_vel;
-		int delta_t;
+		float delta_t;
 		std::uint16_t packet;
-		PlayerHistory(EntityState s, const glm::vec3 &tv, int dt, std::uint16_t p)
+		PlayerHistory(EntityState s, const glm::vec3 &tv, float dt, std::uint16_t p)
 		: state(s), tgt_vel(tv), delta_t(dt), packet(p) { }
 	};
 	std::list<PlayerHistory> player_hist;
